requirements.txt
mysql
mysql-connector
python-wowapi
ratelimit
